% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegGlasser-package.R
\docType{package}
\name{ggsegGlasser-package}
\alias{ggsegGlasser}
\alias{ggsegGlasser-package}
\title{ggsegGlasser: Glasser Atlas for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Glasser multi-modal cortical parcellation (HCP) atlas for the 'ggseg' ecosystem. Provides a unified 'ggseg_atlas' object with 2D polygon geometry for use with 'ggseg'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegGlasser}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegGlasser/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
