% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glasser}
\alias{glasser}
\title{Glasser Multi-Modal Cortical Parcellation (HCP)}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(glasser)
}
\description{
Brain atlas for the Glasser cortical parcellation with 361 regions
based on the Human Connectome Project multi-modal parcellation.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
data(glasser)
glasser
}
\references{
Glasser MF, Coalson TS, Robinson EC, Hacker CD, Harwell J,
Yacoub E, Ugurbil K, Andersson J, Beckmann CF, Jenkinson M, Smith SM,
Van Essen DC (2016). A multi-modal parcellation of human cerebral cortex.
\emph{Nature}, 536(7615):171-178.
\doi{10.1038/nature18933}
}
\concept{ggseg_atlases}
\keyword{datasets}
